fix gravity to be less gas expensive; add gas reports to tests

+++ packages/contracts/src/Constants.sol
@@ -8,7 +8,7 @@ uint32 constant BLOCKS_BEFORE_INCREASE_STAMINA = 60; // 1 minute if 1 block == 1
 uint32 constant STAMINA_INCREASE_RATE = 1666;
 uint16 constant BLOCKS_BEFORE_INCREASE_HEALTH = 60; // 1 minute if 1 block == 1 second
 uint16 constant HEALTH_INCREASE_RATE = 2;
-uint256 constant GRAVITY_DAMAGE = 10;
+uint16 constant GRAVITY_DAMAGE = 10;
 
 int32 constant MAX_PLAYER_BUILD_MINE_HALF_WIDTH = 5;
